Added the latest project versions.
Version "III Mobile" is the working version where support for touch
screens was added. The sound track was added. And a number of bug fixes
/ improvements.

Version "VI" is the shared project on scratch. Currently I'm working on
version III, saving it locally, and then restoring it to version VI on
scratch.mit.edu.